Addons (#46)
* Addon Guide for Lightbar

Add a guide to install the lightbar.

* Add Linear Rail Install Guide

Add installation instructions for the official linear rail kit.

* Add Lightbar Install Tips

Add some tips for the lightbar install to make the cabling clean.

* Add Lightbox Extension

Install the lightbox extension to allow images to be clicked and show up full size in the browser window.

* Instructions for Ceramic Hot End Kit

Add install instructions for the official ceramic hot end kit from Creality.