Updates (#48)
* Update IP Docs

Replace image from IP finding doc with a cleaner version. Update wording on steps.

* Added Tools Recommendations

Added a document on recommended tools when modifying the printer.

* Update MKDocs Dependencies

Update the versions of the mkdocs dependencies.

* Cleanup Nebula Pad Install

Rewrite the install instructions for the Nebula Pad to read cleaner. Also, fix a spelling error on the tools page.

* Hot Glue Tip Revision

Add a tip to soften hot glue with isopyopyl alcohol in addition to tweezers and heat gun methods.

* Update Setup Docs

Add to info about USB drive files. Add tip on formatting the SD card for the firmware.

* Update Tools with Images

Add images to the recommended tools and replace direct product link with a search.